Moved encryption tests to their own test case
